Move unused method out of main code, test symlink
