+++ awx/main/utils/execution_environments.py
@@ -61,17 +61,6 @@ def to_container_path(path, private_data_dir):
     # due to how tempfile.mkstemp works, we are probably passed a resolved path, but unresolved private_data_dir
     resolved_path = Path(path).resolve()
     resolved_pdd = Path(private_data_dir).resolve()
-    if private_data_dir != path and resolved_pdd not in resolved_path.parents:
-        raise RuntimeError(f'Cannot convert path {path} unless it is a subdir of {private_data_dir}')
+    if resolved_pdd != resolved_path and resolved_pdd not in resolved_path.parents:
+        raise RuntimeError(f'Cannot convert path {resolved_path} unless it is a subdir of {resolved_pdd}')
     return str(resolved_path).replace(str(resolved_pdd), CONTAINER_ROOT, 1)
-
-
-def to_host_path(path, private_data_dir):
-    """Given a path inside of the EE container, this gives the absolute path
-    on the host machine within the private_data_dir
-    """
-    if not os.path.isabs(private_data_dir):
-        raise RuntimeError('The private_data_dir path must be absolute')
-    if CONTAINER_ROOT != path and Path(CONTAINER_ROOT) not in Path(path).resolve().parents:
-        raise RuntimeError(f'Cannot convert path {path} unless it is a subdir of {CONTAINER_ROOT}')
-    return path.replace(CONTAINER_ROOT, private_data_dir, 1)
